ccidev: Don't assume CCID interface is 0
Also fix a bug when scanning the interface descriptors where config index would
be off-by-one due to being 1-indexed and not 0-indexed.

This restructures the code to allow us to to locate devices which are either
presenting a standard CCID class interface or from a list of known vendor and
device ID's with a proprietary interface which nevertheless conforms.